+++ src/Symfony/Component/String/CHANGELOG.md
@@ -8,6 +8,8 @@ CHANGELOG
  * made `AbstractString::width()` follow POSIX.1-2001
  * added `LazyString` which provides memoizing stringable objects
  * The component is not marked as `@experimental` anymore.
+ * Added the `s()` helper method to get either an `UnicodeString` or `ByteString` instance,
+   depending of the input string UTF-8 compliancy.
 
 5.0.0
 -----